fix: coin type error and cookie active param

+++ BiliWebApi/src/main/java/com/huanli233/biliwebapi/bean/login_info/CoinCount.kt
@@ -3,5 +3,5 @@ package com.huanli233.biliwebapi.bean.login_info
 import com.google.gson.annotations.Expose
 
 data class CoinCount(
-    @Expose val money: Int
+    @Expose val money: Double
 )
